retrieve runtime input resources e2e test
